+++ src/main/java/com/project/bookuluv/app/admin/controller/ADMController.java
@@ -1,5 +1,6 @@
 package com.project.bookuluv.app.admin.controller;
 
+import com.project.bookuluv.app.article.domain.Article;
 import com.project.bookuluv.app.article.dto.ArticleDto;
 import com.project.bookuluv.app.article.service.ArticleService;
 import com.project.bookuluv.member.domain.Member;
@@ -8,6 +9,7 @@ import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -16,6 +18,7 @@ import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
 import java.security.Principal;
+import java.util.List;
 
 @Controller
 @RequiredArgsConstructor
@@ -41,5 +44,12 @@ public class ADMController {
         Member member = this.memberService.getMember(principal.getName());
         this.articleService.create(articleDto.getSubject(), articleDto.getContent(), member, files);
         return "redirect:/";
+    }
+
+    @GetMapping("/list")
+    public String list(Model model) {
+        List<Article> articleList = this.articleService.getAll();
+        model.addAttribute("articleList", articleList);
+        return "article_list";
     }
 }
